Native Client -The player's position and orientation are now aligned with the camera. -Pressing F1 to shows/hides the bounding volumes of the player's mesh. -Fixed an issue where threads in the asset-loading thread pool would continue to run after the application had been stopped. -Add images so that images would only be loaded once. Common -The transform structure now uses opengl mathematics types for an object's position and orientation. -Single-precision floats are not written and read in binary form.

+++ Common/Include/Starfall/BufferBase.h
@@ -37,11 +37,11 @@ namespace Starfall {
 
 			void writeUInt32(Poco::UInt32 value);
 			void writeString(const string& value);
-			void writeFloat(double value);
+			void writeFloat(float value);
 
 			Poco::UInt32 readUInt32();
 			string		 readString();
-			double		 readFloat();
+			float		 readFloat();
 
 			template<typename T> 
 			void operator>> (T& into);
